Redesign upgrade policies
